pyproject.toml
[tool.poetry]
name = "gpt-arena"
version = "0.1.0"
description = ""
authors = ["David Byttow <david.byttow@gmail.com>"]
readme = "README.md"
packages = [{include = "gpt_arena"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.10.0"
python-dotenv = "^1.0.1"
termcolor = "^2.4.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
